[370] * Added scripts to Rage Winterchill, Anetheron, Kaz'rogal, Azgalor.
* Added scripts to Rage Winterchill, Anetheron, Kaz'rogal, Azgalor.
* Fix Murmur.
* Add proper spells for two of Illidan's spells.
* Fix linux warning in opera script.
* NEW SQL: r369_mangos.sql

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@370 5f9c896b-1e26-0410-94da-f77f675e2462

+++ scripts/zone/karazhan/bosses_opera.cpp
@@ -782,7 +782,7 @@ struct MANGOS_DLL_DECL boss_bigbadwolfAI : public ScriptedAI
     uint32 SwipeTimer;
 
     uint64 HoodGUID;
-    uint32 TempThreat;
+    float TempThreat;
 
     bool IsChasing;
